Fix compiler warnings
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@11571 8aca7d54-2c30-db11-9de9-000461428c89

+++ src/simutrans/builder/wegbauer.cc
@@ -3105,7 +3105,7 @@ void way_builder_t::build_river()
 
 void way_builder_t::build()
 {
-	if(get_count()<2  ||  get_count() > maximum) {
+	if(get_count()<2  ||  get_count() > (uint32)maximum) {
 DBG_MESSAGE("way_builder_t::build()","called, but no valid route.");
 		// no valid route here ...
 		return;
